[refactor] 修改maven group id准备发布到中央仓库
