+++ pom.xml
@@ -2,12 +2,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.shardbatis</groupId>
+    <groupId>com.rookiefly.open.shardbatis</groupId>
     <artifactId>shardbatis</artifactId>
-    <version>2.1.0</version>
+    <version>2.2.0</version>
     <packaging>jar</packaging>
 
-    <name>sharebatis</name>
+    <name>shardbatis</name>
     <url>http://maven.apache.org</url>
 
     <properties>
