+++ src/main/resources/com/rookiefly/open/shardbatis/builder/shardbatis-config.dtd
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+        <!ELEMENT shardingConfig (ignoreList?,parseList?,strategy*)>
+        <!ELEMENT ignoreList (value+)>
+        <!ELEMENT parseList (value+)>
+        <!ELEMENT value (#PCDATA)>
+        <!ELEMENT strategy EMPTY>
+        <!ATTLIST strategy
+                tableName CDATA #REQUIRED
+                strategyClass CDATA #REQUIRED
+                >
