--- src/test/java/com/google/code/shardbatis/test/domain/AppTestDO.java
@@ -1,35 +1,0 @@
-package com.google.code.shardbatis.test.domain;
-
-/**
- * @author sean.he
- */
-public class AppTestDO {
-
-    private int id;
-    private String cnt;
-    private char deleted;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getCnt() {
-        return cnt;
-    }
-
-    public void setCnt(String cnt) {
-        this.cnt = cnt;
-    }
-
-    public char getDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(char deleted) {
-        this.deleted = deleted;
-    }
-}
