+++ src/test/resources/error_config.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE shardingConfig PUBLIC "-//shardbatis.googlecode.com//DTD Shardbatis 2.0//EN"
-        "http://shardbatis.googlecode.com/dtd/shardbatis-config.dtd">
+        "http://open.rookiefly.com/dtd/shardbatis-config.dtd">
 <shardingConfig>
-    <strategy tableName="test_table1" strategyClass="com.google.code.shardbatis.strategy.impl.NoShardStrategy"/>
+    <strategy tableName="test_table1" strategyClass="com.rookiefly.open.shardbatis.strategy.impl.NoShardStrategy"/>
     <strategy tableName="test_table2"/>
 </shardingConfig>
