+++ src/test/resources/test_config_2.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE shardingConfig PUBLIC "-//shardbatis.googlecode.com//DTD Shardbatis 2.0//EN"
-        "http://shardbatis.googlecode.com/dtd/shardbatis-config.dtd">
+        "http://open.rookiefly.com/dtd/shardbatis-config.dtd">
 <shardingConfig>
     <ignoreList>
         <value>ignoreId1</value>
@@ -13,6 +13,6 @@
         <value> <![CDATA[ parseid>2]]></value>
     </parseList>
 
-    <strategy tableName="test_table1" strategyClass="com.google.code.shardbatis.strategy.impl.TestShardStrategyImpl"/>
-    <strategy tableName="test_table2" strategyClass="com.google.code.shardbatis.strategy.impl.TestShardStrategyImpl"/>
+    <strategy tableName="test_table1" strategyClass="com.rookiefly.open.shardbatis.strategy.impl.TestShardStrategyImpl"/>
+    <strategy tableName="test_table2" strategyClass="com.rookiefly.open.shardbatis.strategy.impl.TestShardStrategyImpl"/>
 </shardingConfig>
